reset nested changes; added test file to track
